requirements.txt
PyQt6==6.7.1
PyOpenGl==3.1.7
Numpy==2.1.1
SciPy==1.14.1
GLFW==2.7.0
PyInstaller==6.10.0